LOK-1891: Default to introspection disabled
